#EXT-X-TARGETDURATION:<s> is supposed to be a decimal-integer
https://datatracker.ietf.org/doc/html/rfc8216#section-4.3.3.1

+++ src/playlist.rs
@@ -713,7 +713,7 @@ impl SessionData {
 pub struct MediaPlaylist {
     pub version: Option<usize>,
     /// `#EXT-X-TARGETDURATION:<s>`
-    pub target_duration: f32,
+    pub target_duration: u64,
     /// `#EXT-X-MEDIA-SEQUENCE:<number>`
     pub media_sequence: u64,
     pub segments: Vec<MediaSegment>,
